+++ src/gui.h
@@ -25,7 +25,7 @@ public:
 
 private:
     void renderMenu();
-    bool LoadTextureFromMemory(const void *data, unsigned long long data_size, GLuint *out_texture, int *out_width, int *out_height);
+    bool LoadTextureFromMemory(const void *data, uint64_t data_size, GLuint *out_texture, int *out_width, int *out_height);
     bool LoadTextureFromFile(const char *file_name, GLuint *out_texture, int *out_width, int *out_height);
 
 public:
